test(ui): share composer jsdom harness (CODE-164)
